refactor(phase-3k): extract JiraIssueService (issue domain)
Phase 3k of ADR-002 — the final big slice of the jira_client.py
decomposition. The five issue-domain methods move into a focused
``JiraIssueService`` exposed on ``JiraClient`` as ``self.issues``:

- ``get_all_issues_for_project`` (full-project pagination)
- ``get_issue_details`` (single-issue detail extraction)
- ``batch_get_issues`` (JQL-based batch fetch)
- ``_fetch_issues_batch`` (private helper, removed from JiraClient)
- ``stream_all_issues_for_project`` (memory-efficient streaming
  paginator with ``@rate_limited()``)

Each public method keeps a thin delegator on ``JiraClient`` so existing
call sites continue to work unchanged.

Cross-service callers
---------------------

``get_all_issues_for_project`` is also called from:

- ``JiraTempoService`` (no direct call — it uses Tempo HTTP endpoints)
- ``JiraWorklogService.get_work_logs_for_project`` via
  ``self._client.get_all_issues_for_project(...)``

Those call sites keep working because the delegator stays in place on
``JiraClient``. No changes to those services.

Cleanup
-------

- Drops the now-unused ``StreamingPaginator`` and ``rate_limited``
  imports from ``jira_client.py`` (they only existed to support the
  extracted ``stream_all_issues_for_project``).
- ``EnhancedJiraClient`` already overrides ``batch_get_issues`` /
  ``_fetch_issues_batch`` with its own parallel implementations, so
  removing the private helper from ``JiraClient`` does not affect the
  enhanced-client tests.

LOC: jira_client.py 1009 → 812 (-197).

Verification
------------

  ruff check   src/clients/jira_client.py src/clients/jira_issue_service.py  -> clean
  ruff format  src/clients/jira_client.py src/clients/jira_issue_service.py  -> unchanged
  mypy src/                                                                  -> Success
  pytest tests/unit -x -q                                                    -> 953 passed
